Removing the need for bootstrapping
For this role specifically bootstrapping doen't make sense as this would create a controller, but this is the role that would do that.

Instead we login as the vagrant user directly (when running locally), which we can already do through the provisioning steps performed in the vagrantfile.

This removes the need for us to depend on any roles, which matches the dependencies for this role generally.