Exlcuding testing a user account with password from CI tests
The hasing function used when setting a password for a user generates a random salt. This means each time the task is run a change event occurs.

This breaks the idempotency test. Given this feature is not recommended I am excluding this task from the CI tests (this is still tested by the local tests). We could use a fixed hash but that's less secure so want to avoid that.