Adding ability to add users to sudo groups
This removes the need for users to work out which sudo group a user needs to be added to using a group_var for each OS for example.

The logic for this is fairly simple since Ansible knows the OS distribution through facts we just use a value per OS, and then use that value if a user should be a sudo user.