Switching to secondary group append mode
This means where secondary groups are specified, they are appended to any existing secondary groups, rather than replacing them.

This was needed to fix issues where the 'groups' task file would set the generic secondary groups for a user, and the 'sudo' task file would then overwrite this if it added the user to the relevant sudo group.

Using append mode for just the sudo tasks would not work as the groups tasks would reset this on each play, breaking idempotency.

This represents a breaking change as users may have relied on this previously behaviour.