Removing the need to always specify authorized_keys_files
Previously this option had to be defined, even if no authorised keys were defined, because of its use in a 'with_subelements' loop.

This meant authorized_keys_directory was used as a proxy to determine if keys should actually be managed or not.

Now authorised files is optional, this property can be used directly for this purpose.

This means authorised directory can be removed, with authorised files now containing the entire path to each key (which removes some unnecessary complexity.